Refactor XFCC to use a stand-alone marshaller
